+++ engine/src/hopeit/server/config.py
@@ -6,8 +6,6 @@ from enum import Enum
 from typing import TypeVar, List, Optional
 import re
 import os
-from functools import partial
-from pydantic import SecretStr
 
 from hopeit.dataobjects import dataclass, dataobject, field
 from hopeit.dataobjects.payload import Payload
@@ -40,8 +38,6 @@ class StreamsConfig:
     :stream_manager: str: Stream manager class name. Default is "hopeit.streams.NoStreamManager".
     :field connection_str: str, url to connect to streams server: i.e. redis://localhost:6379
         if using redis stream manager plugin to connect locally
-    :field username: SecretStr: Username for authentication. Default is an empty secret string.
-    :field password: SecretStr: Password for authentication. Default is an empty secret string.
     :field delay_auto_start_seconds: int: Delay in seconds before auto-starting the stream.
         Default is 3 seconds.
     :field initial_backoff_seconds: float: Initial backoff time in seconds for connection retries.
@@ -57,8 +53,6 @@ class StreamsConfig:
 
     stream_manager: str = "hopeit.streams.NoStreamManager"
     connection_str: str = "<<NoStreamManager>>"
-    username: SecretStr = field(default_factory=partial(SecretStr, ""))
-    password: SecretStr = field(default_factory=partial(SecretStr, ""))
     delay_auto_start_seconds: int = 3
     initial_backoff_seconds: float = 1.0
     max_backoff_seconds: float = 60.0