+++ plugins/data/dataframes/src/hopeit/dataframes/serialization/files.py
@@ -70,7 +70,7 @@ class DatasetFileStorage(Generic[DataFrameT]):
         """
         datatype: Type[DataFrameT] = type(dataframe)
         return await self.save_df(
-            dataframe._df.select(datatype.__dataframe__.columns),
+            dataframe._df.select(datatype.__dataframe__.columns),  # type: ignore[misc]
             datatype,
             partition_dt=partition_dt,
             database_key=database_key,